refactored base and improved type hinting
