+++ digitalpy/core/IAM/persistence/__init__.py
@@ -3,9 +3,6 @@ IAM persistence module
 """
 # pylint: disable=unused-import
 # pylint: disable=wrong-import-position
-from sqlalchemy.ext.declarative import declarative_base
-
-IAMBase = declarative_base()
 
 from .session_contact import SessionContact
 from .system_group_permission import SystemGroupPermission
