+++ digitalpy/core/IAM/persistence/iam_base.py
@@ -1,0 +1,3 @@
+from sqlalchemy.ext.declarative import declarative_base
+
+IAMBase = declarative_base()
